analytics/requirements.txt
openpyxl
pandas
faker
ipykernel
python-dotenv
requests
networkx
matplotlib
